+++ spec/sidekiq/rescue/dsl_spec.rb
@@ -103,7 +103,20 @@ RSpec.describe Sidekiq::Rescue::Dsl do
     it "raises ArgumentError if delay is not an integer or float" do
       expect { define_dsl { sidekiq_rescue TestError, delay: "60" } }.to raise_error(
         ArgumentError,
-        "delay must be integer, float or proc"
+        "delay must be integer, float, proc, or one of: polynomially_longer"
+      )
+    end
+
+    it "sets the delay symbol :polynomially_longer" do
+      define_dsl { sidekiq_rescue TestError, delay: :polynomially_longer }
+
+      expect(job_class.sidekiq_rescue_options.dig([TestError], :delay)).to eq(:polynomially_longer)
+    end
+
+    it "raises ArgumentError if delay is an unrecognized symbol" do
+      expect { define_dsl { sidekiq_rescue TestError, delay: :foo } }.to raise_error(
+        ArgumentError,
+        "delay must be integer, float, proc, or one of: polynomially_longer"
       )
     end